+++ packages/core/src/definitions/cls.ts
@@ -21,6 +21,10 @@ export type ConstructorArgsDefinitions<T extends any[], TCurrentLifeTime extends
   [K in keyof T]: Definition<T[K] | Promise<T[K]>, ValidDependenciesLifeTime<TCurrentLifeTime>, any>;
 };
 
+export type Callable<TInstance, TArgs extends any[]> = {
+  call(...args: TArgs): TInstance;
+};
+
 export const cls = {
   transient: <
     TInstance,