+++ packages/core/src/definitions/impl/ClassDefinition.ts
@@ -1,7 +1,7 @@
 import type { ClassType } from '../cls.js';
 import { type Thunk, unwrapThunk } from '../../utils/Thunk.js';
 import type { IContainer, IStrategyAware } from '../../container/IContainer.js';
-import { isPromise } from '../../utils/IsPromise.js';
+import { isThenable } from '../../utils/IsThenable.js';
 import type { LifeTime } from '../abstract/LifeTime.js';
 import type { InstancesDefinitions } from '../abstract/InstanceDefinition.js';
 import type { AnyDefinition, IDefinition } from '../abstract/IDefinition.js';
@@ -16,11 +16,11 @@ export class ClassDefinition<TInstance, TLifeTime extends LifeTime, TConstructor
   constructor(
     public readonly id: symbol,
     public readonly strategy: TLifeTime,
-    public readonly klass: ClassType<TInstance, TConstructorArgs>,
-    public readonly dependencies?: Thunk<InstancesDefinitions<TConstructorArgs, TLifeTime>>,
+    protected readonly _class: ClassType<TInstance, TConstructorArgs>,
+    protected readonly _dependencies?: Thunk<InstancesDefinitions<TConstructorArgs, TLifeTime>>,
   ) {
-    if (Array.isArray(dependencies)) {
-      this.assertValidDependencies(dependencies);
+    if (Array.isArray(_dependencies)) {
+      this.assertValidDependencies(_dependencies);
     }
   }
 
@@ -36,30 +36,30 @@ export class ClassDefinition<TInstance, TLifeTime extends LifeTime, TConstructor
 
   create(use: IContainer): TInstance {
     // no dependencies
-    if (this.dependencies === undefined) {
+    if (this._dependencies === undefined) {
       // @ts-ignore
-      return new this.klass();
+      return new this._class();
     }
 
-    const deps = use.all(...unwrapThunk(this.dependencies)) as TConstructorArgs | Promise<TConstructorArgs>;
+    const deps = use.all(...unwrapThunk(this._dependencies)) as TConstructorArgs | Promise<TConstructorArgs>;
 
     if (this._hasOnlySyncDependencies) {
-      return new this.klass(...(deps as TConstructorArgs));
+      return new this._class(...(deps as TConstructorArgs));
     }
 
-    if (isPromise(deps)) {
+    if (isThenable(deps)) {
       return deps.then(deps => {
-        return new this.klass(...deps);
+        return new this._class(...deps);
       }) as TInstance;
     } else {
       this._hasOnlySyncDependencies = true;
 
-      return new this.klass(...deps);
+      return new this._class(...deps);
     }
   }
 
   get name() {
-    return this.klass.name;
+    return this._class.name;
   }
 
   private assertValidDependencies(dependencies: AnyDefinition[]) {