+++ packages/core/src/patching/__test__/decorate.test.ts
@@ -12,7 +12,7 @@ describe(`decorate`, () => {
     const someValue = value(1);
 
     const c = container.new(c => {
-      c.bind(someValue).decorate((_, val) => val + 1);
+      c.bind(someValue).toDecorated(val => val + 1);
     });
 
     expect(c.use(someValue)).toEqual(2);
@@ -24,10 +24,41 @@ describe(`decorate`, () => {
     expect(container.new().use(someValue)).toEqual(1);
 
     const cnt = container.new(c => {
-      c.bind(someValue).decorate((_, val) => val + 1);
+      c.bind(someValue).toDecorated(val => val + 1);
     });
 
     expect(cnt.use(someValue)).toEqual(2);
+  });
+
+  it(`calls decorate function with awaited instance in case of async definitions`, async () => {
+    const someValue = fn(async () => 1);
+
+    const c = container.new(c => {
+      c.bind(someValue).toDecorated(val => val + 1);
+    });
+
+    expect(await c.call(someValue)).toEqual(2);
+  });
+
+  it(`allows using async decorate Fn for async definitions`, async () => {
+    const someValue = fn(async () => 1);
+
+    const c = container.new(c => {
+      c.bind(someValue).toDecorated(async val => val + 1);
+    });
+
+    expect(await c.call(someValue)).toEqual(2);
+  });
+
+  it(`does not allow async decorate functions for sync definitions`, async () => {
+    const someValue = value(1);
+
+    const c = container.new(c => {
+      //@ts-expect-error - cannot return Promise<number> for number (due to async function)
+      c.bind(someValue).toDecorated(async val => val + 1);
+    });
+
+    expect(() => c.use(someValue)).toThrowError();
   });
 
   it(`allows using additional dependencies, ex1`, async () => {
@@ -36,7 +67,7 @@ describe(`decorate`, () => {
     const someValue = value(10);
 
     const c = container.new(c => {
-      c.bind(someValue).decorate((use, val) => {
+      c.bind(someValue).toDecorated((val, use) => {
         const aVal = use(a);
         const bVal = use(b);
 
@@ -52,11 +83,11 @@ describe(`decorate`, () => {
     const b = value(2);
 
     const someValue = fn.scoped(use => {
-      return use(a) + use(b);
+      return use.call(a) + use.call(b);
     });
 
     const c = container.new(c => {
-      c.bind(someValue).decorate((use, val) => {
+      c.bind(someValue).toDecorated((val, use) => {
         return val * use(b);
       });
     });
@@ -69,7 +100,7 @@ describe(`decorate`, () => {
       const a = fn.scoped(() => Math.random());
 
       const c = container.new(c => {
-        c.bind(a).decorate((use, a) => a);
+        c.bind(a).toDecorated(a => a);
       });
 
       expect(c.use(a)).toEqual(c.use(a));
@@ -79,10 +110,10 @@ describe(`decorate`, () => {
       const a = fn(() => Math.random());
 
       const c = container.new(c => {
-        c.bind(a).decorate((use, a) => a);
+        c.bind(a).toDecorated(a => a);
       });
 
-      expect(c.use(a)).not.toEqual(c.use(a));
+      expect(c.call(a)).not.toEqual(c.call(a));
     });
 
     it(`uses correct scope`, async () => {
@@ -93,7 +124,7 @@ describe(`decorate`, () => {
       });
 
       const c = container.new(c => {
-        c.bind(a).decorate((use, a) => a);
+        c.bind(a).toDecorated(a => a);
       });
 
       const obj1 = fn.scoped(use => ({
@@ -122,10 +153,8 @@ describe(`decorate`, () => {
   describe(`globalOverrides`, () => {
     function setup(instanceDef: Definition<MyService, ScopeConfigureAllowedLifeTimes, []>) {
       const scope1 = container.new(c => {
-        c.freeze(instanceDef).configure((use, a) => {
+        c.freeze(instanceDef).toConfigured(a => {
           vi.spyOn(a, 'callMe');
-
-          return a;
         });
       });