+++ packages/core/src/utils/IsPromise.ts
@@ -1,3 +1,0 @@
-export function isPromise<T>(value: any): value is Promise<T> {
-  return typeof value?.then === 'function';
-}